backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
openai
pinecone-client
httpx
googlemaps
slowapi
pytest
sqlalchemy
bcrypt
python-jose[cryptography]
python-multipart
email-validator
psycopg2-binary